+++ internal/predictor/constants.go
@@ -22,7 +22,8 @@ package predictor
 var AllListFields = map[string]string{
 	"Id":            "Column Index",
 	"Arn":           "AWS Role Resource Name",
-	"AccountId":     "AWS AccountID",
+	"AccountId":     "AWS AccountID (integer)",
+	"AccountIdStr":  "AWS AccountID (zero padded)",
 	"AccountName":   "Configured Account Name",
 	"AccountAlias":  "AWS Account Alias",
 	"DefaultRegion": "Default AWS Region",